Shifted registers viewsets to api_v2 file and disable Changerequest api create and update functionality
